feat: store nonexistent devices and places as null
According to the new scheme, null value means that the resource has been
fetched and it's missing in the database, whereas undefined means that
the fetching hasn't been initiated yet.

It perfectly follows the approach of serving Place.device_id:
- it's undefined if it hasn't been requested manually
- it's null if the device isn't bound to the given place
- it's a device ID if the device is bound to the given place.

This approach provides the ability to differenciate between the
resources that haven't been requested yet (=== undefined) and those ones
that are missing in the database (=== null).